backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email-validator]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
ultralytics==8.2.100
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
prophet==1.1.6
pandas==2.2.3
